Create Hint component and fix styling
