+++ src/routes/kryzz/[slug]/+page.svelte
@@ -1,12 +1,10 @@
 <script lang="ts">
-	import { Confetti } from 'svelte-confetti';
 	import '@oddbird/popover-polyfill';
 
 	import type { PageData } from './$types';
 	import { createCrosswordContext } from '$lib/contexts/CrosswordContext.svelte';
-	import { enhance } from '$app/forms';
-	import type { SubmitFunction } from '@sveltejs/kit';
 	import CrosswordBoard from '$lib/components/crossword/CrosswordBoard.svelte';
+	import Hint from '$lib/components/crossword/Hint.svelte';
 
 	type Props = {
 		data: PageData;
@@ -15,83 +13,25 @@
 	let { data }: Props = $props();
 
 	const crossword = $state(data.crossword);
-	let isCorrect: boolean | undefined = $state();
-	let showMessage = $state(false);
 
-	const crosswordContext = createCrosswordContext(crossword);
+	createCrosswordContext(crossword);
 
 	const keydown = (event: KeyboardEvent) => {
 		if (event.metaKey || event.key === 'Shift') {
 			return;
 		}
 	};
-
-	const onCheckSolve: SubmitFunction = () => {
-		if (!crosswordContext.isFull) {
-			showMessage = true;
-			return;
-		}
-		return ({ result }) => {
-			if (result.type === 'success') {
-				isCorrect = result?.data?.correct;
-			}
-		};
-	};
-
-	let showError = $derived(isCorrect === false);
-
-	const windowClickHandler = () => {
-		isCorrect = undefined;
-		showMessage = false;
-	};
 </script>
 
-<svelte:window onkeydown={keydown} onclick={windowClickHandler} />
+<svelte:window onkeydown={keydown} />
 
 <div class="crossword">
 	<h2>{crossword?.name}</h2>
 	<div class="crossword-container">
 		<CrosswordBoard />
-		<div class="hint">
-			{#if showError}
-				<p class="error">Nope, inte riktigt rätt än</p>
-			{/if}
-			{#if showMessage}
-				<p class="message">Du måste fylla i hela kryzzet först</p>
-			{/if}
-		</div>
-		<div class="form-buttons">
-			<form method="POST" action="?/submit" use:enhance={onCheckSolve} style="position: relative;">
-				<input type="hidden" name="crosswordId" value={crossword.crosswordId} />
-				<button>Rätta</button>
-				{#if isCorrect}
-					<div class="confetti">
-						<Confetti infinite y={[1, 2]} x={[-1, 1]} amount={100} cone />
-					</div>
-				{/if}
-			</form>
-			<button
-				class="danger-button"
-				type="button"
-				popovertarget="deleteDialog"
-				popovertargetaction="show">Rensa</button
-			>
-		</div>
+		<Hint />
 	</div>
 </div>
-
-<dialog id="deleteDialog" popover>
-	<form method="POST" action="?/reset">
-		<p>Vill du verkligen rensa hela kryzzet?</p>
-		<input type="hidden" name="crosswordId" value={crossword.crosswordId} />
-		<div class="buttons">
-			<button type="button" popovertarget="deleteDialog" popovertargetaction="hide">Avbryt</button>
-			<button class="danger-button" popovertarget="deleteDialog" popovertargetaction="hide"
-				>Rensa</button
-			>
-		</div>
-	</form>
-</dialog>
 
 <style>
 	.crossword {
@@ -100,21 +40,6 @@
 		display: flex;
 		flex-direction: column;
 		align-items: center;
-	}
-
-	h2 {
-		margin-top: 0;
-		font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
-		font-weight: 200;
-		letter-spacing: 0.1em;
-		color: #222;
-		font-size: 1.5rem;
-	}
-
-	@media (min-width: 46.25rem) {
-		h2 {
-			font-size: 2rem;
-		}
 	}
 
 	.crossword-container {
@@ -131,99 +56,7 @@
 		}
 	}
 
-	dialog p {
-		font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
-	}
-
-	[popover] {
-		max-width: 32em;
-		border-radius: 0.25em;
-		border: none;
-		padding: 1em;
-	}
-
-	[popover]::backdrop {
-		background: rgba(0, 0, 0, 0.3);
-	}
-
-	[popover]:popover-open {
-		animation: intro 0.3s cubic-bezier(0.34, 1.56, 0.64, 1);
-	}
-
-	@keyframes intro {
-		from {
-			transform: scale(0.95);
-		}
-		to {
-			transform: scale(1);
-		}
-	}
-
-	[popover]:popover-open::backdrop {
-		animation: outro 0.2s ease-out;
-	}
-
-	@keyframes outro {
-		from {
-			opacity: 0;
-		}
-		to {
-			opacity: 1;
-		}
-	}
-
-	#deleteDialog .buttons {
-		display: flex;
-		justify-content: flex-end;
-		gap: 1rem;
-	}
-
-	button {
-		border: none;
-		height: 36px;
-		min-width: 88px;
-		color: rgba(0, 0, 0, 0.78);
-		font-size: 14px;
-		text-transform: uppercase;
-		cursor: pointer;
-		padding: 0 16px;
-		border-radius: 1px;
-
-		letter-spacing: 0.01em;
-		font-weight: 500;
-		transition:
-			box-shadow 0.4s cubic-bezier(0.25, 0.8, 0.25, 1),
-			background-color 0.4s cubic-bezier(0.25, 0.8, 0.25, 1);
-	}
-
-	button.danger-button {
-		color: red;
-	}
-
-	.confetti {
-		position: absolute;
-		top: 50%;
-		left: 50%;
-		pointer-events: none;
-	}
-
-	.form-buttons {
-		width: 100%;
-		display: flex;
-		justify-content: space-between;
-	}
-
-	.hint {
-		min-height: 24px;
-		margin-block: 1rem;
-	}
-
-	.hint > p {
-		margin: 0;
-		text-align: center;
-	}
-
-	.error {
-		color: red;
+	h2 {
+		margin-top: 0;
 	}
 </style>
